moved remaining resolvers and routes from user entities in /server/xpub-server
