+++ packages/component-elife-app/server/routes/publicConfig.js
@@ -1,5 +1,7 @@
 const config = require('config')
 const { pickBy } = require('lodash')
+const logger = require('@pubsweet/logger')
+const jwt = require('jsonwebtoken')
 
 const template = clientConfig => `
   window.config = ${JSON.stringify(clientConfig)}
@@ -12,4 +14,19 @@ module.exports = app => {
     res.type('js')
     res.send(response)
   })
+
+  if (config.get('login.enableMock')) {
+    app.get('/mock-token-exchange/:id', (req, res) => {
+      const secret = config.get('pubsweet-server.secret')
+      const mockProfileId = req.params.id
+      logger.info('Signing mock JWT for profile ID', mockProfileId)
+      const xpubToken = jwt.sign(
+        { id: mockProfileId, iss: 'journal', 'new-session': true },
+        secret,
+        { expiresIn: '1m' },
+      )
+
+      res.redirect(`/login#${xpubToken}`)
+    })
+  }
 }
