+++ server/xpub-server/index.js
@@ -1,5 +1,4 @@
 const os = require('os')
-const { merge } = require('lodash')
 const fs = require('fs')
 const logger = require('@pubsweet/logger')
 const cleanup = require('./cleanup')
@@ -13,20 +12,13 @@ const elifeTypeDefs = fs.readFileSync(
   `${__dirname}/schema/elife.graphqls`,
   'utf8',
 )
-const entityTypeDefs = [
-  fs.readFileSync(`${__dirname}/entities/user/typeDefs.graphqls`, 'utf8'),
-]
+
 const typeDefs = `
   ${xpubTypeDefs}
   ${elifeTypeDefs}
-  ${entityTypeDefs.join('\n\n')}
 `
 
-// merge resolvers
-const resolvers = merge({}, require(`./entities/user/resolvers`))
-
 const registerRoutes = app => {
-  require(`./entities/user/routes`)(app)
   require('./routes')(app)
 }
 
@@ -46,6 +38,5 @@ const cleanupHandler = cleanup.Cleanup(process, logger, stopServer)
 module.exports = {
   backend: () => registerRoutes,
   typeDefs,
-  resolvers,
   migrationsPath: `./schema/migrations`,
 }
